Update ChocoboColor.java
The purpose of this (personal) fork is rebalancing high tier chocobos a little bit for a private modpack, particularly fly speed makes other flying mods extremely obsolete. Might take a few tries to see how this works out.

I'm not too familiar with Git but i *think* this is a personal fork now so it shouldn't tie in to the original at all..I hope.